services/ai-automation-service/requirements.txt
# AI Automation Service Dependencies
# Phase 1 MVP - Keep it simple

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# MQTT
paho-mqtt==1.6.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# InfluxDB Client
influxdb-client==1.38.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# LLM Integration - OpenAI (Optional, for high-quality suggestions)
openai==1.12.0

# Phase 1 Optimized Model Stack (OpenVINO INT8)
# Version constraints to avoid conflicts:
# optimum-intel 1.20.0 requires transformers<4.46 and >=4.36
# CPU-only PyTorch index (no CUDA - saves 8.5GB!)
--extra-index-url https://download.pytorch.org/whl/cpu

sentence-transformers==3.3.1  # Embeddings: all-MiniLM-L6-v2
transformers==4.45.2          # HuggingFace models (compatible with optimum-intel)
torch                         # PyTorch CPU-only (1.5GB vs 10GB with CUDA)
openvino==2024.5.0            # Intel optimization
optimum-intel==1.20.0         # OpenVINO integration with Transformers

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

